refactor: 샤드 아이디 생성기에 비트 관리를 IdComponent에 위임

+++ src/main/java/ShardIdAllocator.java
@@ -1,4 +1,4 @@
 @FunctionalInterface
 public interface ShardIdAllocator {
-    long allocate(final int maxAllowedShardId);
+    long allocate(final long maxAllowedShardId);
 }
